Move constants.py to _internal
