+++ certbot-nginx/certbot_nginx/_internal/configurator.py
@@ -24,7 +24,7 @@ from certbot import util
 from certbot.compat import os
 from certbot.plugins import common
 
-from certbot_nginx import constants
+from certbot_nginx._internal import constants
 from certbot_nginx import display_ops
 from certbot_nginx import http_01
 from certbot_nginx import nginxparser
